nats: Split architecture specific code out
Move architecture specific code out into its own file ready for enabling ARM
NATS supports.
